Added functionalities and error handeling, and removed bugs

+++ inc/lightSensor.h
@@ -41,8 +41,8 @@ typedef struct Light
 	int IsError;
 }light_t;
 
-uint8_t startup_test(void);
-uint8_t power_up(void);
+int startup_test(void);
+int power_up(void);
 uint8_t read_control_register(void);
 uint8_t set_timing_register(uint8_t integ, uint8_t gain);
 uint8_t read_timing_register(void);
