+++ inc/temperature.h
@@ -30,7 +30,7 @@ tempt_t read_temperature(void);
 void cal_temp(int16_t temperature_hex, float *celcius, float *farenheit, float *kelvin);
 float read_Tlow(void);
 float read_Thigh(void);
-uint16_t read_configuration_reg(void);
+int read_configuration_reg(void);
 void set_shutdown(void);
 void disable_shutdown(void);
 uint8_t read_fault(void);
